+++ integration_test/disclosure_session/filled_app_scenarios/filled_choice.dart
@@ -5,6 +5,7 @@ import 'package:irmamobile/src/screens/session/disclosure/widgets/disclosure_per
 import 'package:irmamobile/src/screens/session/disclosure/widgets/disclosure_permission_make_choice_screen.dart';
 import 'package:irmamobile/src/widgets/credential_card/irma_credential_card.dart';
 import 'package:irmamobile/src/widgets/irma_card.dart';
+import 'package:irmamobile/src/widgets/requestor_header.dart';
 
 import '../../helpers/helpers.dart';
 import '../../helpers/issuance_helpers.dart';
@@ -166,6 +167,14 @@ Future<void> filledChoiceTest(WidgetTester tester, IntegrationTestIrmaBinding ir
   // Confirm choice
   await tester.tapAndSettle(find.text('Done'));
   expect(find.byType(DisclosurePermissionChoicesScreen), findsOneWidget);
+
+  final requestorHeaderFinder = find.byType(RequestorHeader);
+  await evaluateRequestorHeader(
+    tester,
+    requestorHeaderFinder,
+    localizedRequestorName: 'demo.privacybydesign.foundation',
+    isVerified: false,
+  );
 
   // Only selected card should remain
   await evaluateCredentialCard(